backend/requirements.txt
flask>=3.0
flask-sock>=0.7
simple-websocket>=1.0
pyelftools>=0.30
capstone>=5.0
cxxfilt>=0.3
